# swagger.json
{
	"components": {
...
		"schemas": {
			"IErrorMessage": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EEntityStatus": {
				"enum": [
					"ACTIVE",
					"DISABLED"
				],
				"type": "string"
			},
			"ECompanyPlanType": {
				"enum": [
					"BASIC",
					"INTERMEDIATE",
					"ADVANCED"
				],
				"type": "string"
			},
			"EOrderType": {
				"enum": [
					"DESC",
					"ASC"
				],
				"type": "string"
			},
			"HttpSuccessStatusCode": {
				"enum": [
					200,
					201,
					204
				],
				"type": "number"
			},
...
			"CreatedResult": {
				"properties": {
					"data": {},
					"message": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/HttpSuccessStatusCode"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NoContentResult": {
				"properties": {
					"data": {},
					"message": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/HttpSuccessStatusCode"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EVehicleType": {
				"enum": [
					0,
					1,
					2
				],
				"type": "number"
			},
			"CreateVehicleDto": {
				"properties": {
					"plate": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EVehicleType"
					},
					"capacity": {
						"type": "string"
					},
					"driverId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"plate",
					"model",
					"type",
					"driverId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_UpdateVehicleDto.Exclude_keyofUpdateVehicleDto.id__": {
				"properties": {
					"plate": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EVehicleType"
					},
					"capacity": {
						"type": "string"
					},
					"driverId": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_UpdateVehicleDto.id_": {
				"$ref": "#/components/schemas/Pick_UpdateVehicleDto.Exclude_keyofUpdateVehicleDto.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "shippify-technical-test-server",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/vehicle": {
			"get": {
				"operationId": "ListVehicles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					}
				},
				"summary": "List vehicles",
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "plate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "model",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/EVehicleType"
						}
					},
					{
						"in": "query",
						"name": "capacity",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/EEntityStatus"
						}
					},
					{
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "driverFirstName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "driverLastName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "driverEmail",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "driverStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/EEntityStatus"
						}
					},
					{
						"in": "query",
						"name": "companyId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "companyCity",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "companyStatus",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/EEntityStatus"
						}
					},
					{
						"in": "query",
						"name": "companyPlanType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ECompanyPlanType"
						}
					},
					{
						"in": "query",
						"name": "take",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "orderType",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/EOrderType"
						}
					},
					{
						"in": "query",
						"name": "activeOnly",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateVehicle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreatedResult"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					}
				},
				"summary": "Creates a new vehicle",
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVehicleDto"
							}
						}
					}
				}
			}
		},
		"/vehicle/{id}": {
			"get": {
				"operationId": "GetVehicleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					}
				},
				"summary": "Find vehicle by id",
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "driver",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "company",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "activeOnly",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateVehicle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContentResult"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					}
				},
				"summary": "Updates a vehicle by it's id",
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_UpdateVehicleDto.id_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DisableVehicle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"204": {
						"description": ""
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorMessage"
								}
							}
						}
					}
				},
				"summary": "Disable a vehicle by it's id",
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}